refactor: サイド関連コンポーネントを新設したサイドフォルダに集約
