+++ src/app/(authenticated)/_components/side/SideArea.tsx
